+++ SKO_NPC.h
@@ -9,23 +9,23 @@ class SKO_NPC
 {
 
 public:
-	  SKO_NPC();
-      void Respawn();
-      
-      //coords
-      float x, y;
-      int x1,x2,y1,y2;
+	SKO_NPC();
+	void Respawn();
+
+	//coords
+	float x, y;
+	int x1,x2,y1,y2;
 	float x_speed, y_speed;
-int w, h;
-      //spawn
-      int sx, sy;
-     int AI_ticker;       
-	int AI_period;
-       
-      bool facing_right; 
-      bool ground;
-      int current_frame;
-      int animation_ticker;
+	int w, h;
+	//spawn
+	int sx, sy;
+	unsigned long int AI_ticker;       
+	unsigned long int AI_period;
+
+	bool facing_right; 
+	bool ground;
+	int current_frame;
+	unsigned long int animation_ticker;
